Move previous config management code to separate package
